Fix the build-system, internalizing all the deps
This will need to be changed a bit more when we get around to packaging
the 2.0 release on AUR.

In more detail, we currently link dynamically to a local version of the
dependent libraries, most major of which being vmime.

This has the benefit of making it very simple to build the project
locally and contribute, regardless of the base operating system. However
in packaging it creates problems. Some example solutions on how to fix
this are:
a) build the executable statically (this creates a pretty large file)
b) make all deps dynamic on AUR, and package missing libraries.
c) don't bother with packaging them, install shared libraries alongside
  in some /lib/ folder of the package (as required artifacts, i.e. a
  non-shared shared library). This is +- what we do now, w/o installing.
d) all of the above, link some deps one way and others the other way

Signed-off-by: Tomasz Kulis <pub@tfkls.dev>